refactor: isolate cache storage
