+++ docs/superpowers/plans/2026-08-17-modularization-plan.md
@@ -66,7 +66,7 @@
 **Files:** `lib/MeshCache.lua`, new `lib/CacheStorage.lua`, new `lib/CacheIdentity.lua`, new `lib/CacheManifest.lua`, cache tests
 
 - [x] Characterize cache file names, metadata, compression, validation, fingerprints, and incomplete-build resume records.
-- [ ] Move filesystem/codec operations into `CacheStorage`.
+- [x] Move scoped-storage and byte/table fallback operations into `CacheStorage`.
 - [x] Move build identity and invalidation rules into `CacheIdentity`.
 - [x] Move manifest and resume-record handling into `CacheManifest`.
 - [x] Keep `MeshCache` as the stable façade and compare serialized outputs against the baseline fixtures.
